feat: integrate Paystack bank selection and update property creation flow
